pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mac-access-api"
version = "0.1.0"
description = "Secure always-on macOS remote control API with MCP streamable HTTP"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic-settings>=2.10.0",
  "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "httpx>=0.28.0"
]

[tool.setuptools.packages.find]
include = ["mac_access_api*"]

[tool.pytest.ini_options]
pythonpath = ["."]
